ENH RFC renaming, clf criteria and memory optim
ENH Added classification axis criteria to work with multioutput bipartite adapter.

RFC Multiple renaming of bipartite criterion adapters

RFC X_rows, X_cols and y.T are no more created as copies of X/y and stored by the BipartiteSplitter, since this was greatly increasing memory usage of forests (each tree was holding a full copy of the dataset!).

RFC A different version of the multioutput bipartite adapter (currently called PBCTCriterionWrapper but soon to be renamed) was implemented using the new `total_node_impurity()` methods of AxisCriterion. It was made to, when `prediction_weights='leaf_uniform`, avoid storing an output value for each of the (n_rows + n_cols) samples in the dataset in each node, since only the node's average was going to be used anyway. The old method is still necessary for the other prediction_weights options of tree-based models, and still generates a much larger tree object: the node value for each node is a (n_rows+n_cols)-sized vector with NaNs for samples not in the node.

Options I see to tackle that issue are 1. to only store the node value of leaves (which is done to allow inspection, according to comments in _tree.pyx, and I also recognize it facilitates prunning); 2. store sample_indices in the Tree and the start/end positions in each node, which could enable avoiding the NaNs in the node values. However, since node values are stored in a matrix by the Tree object, a different number of outputs for each node would require a ragged array to achieve any improvement in memory usage. Any of the options are hindered by the unflexible structure of the Tree objects.

+++ tests/make_examples.py
@@ -188,7 +188,7 @@ def make_interaction_regression(
         for s, f in zip(n_samples, n_features)
     ]
     X_molten = row_cartesian_product(X)
-    Y_molten = tree.predict(X_molten)
+    Y_molten = tree.predict(X_molten).astype(np.float64)
     if noise > 0.0:
         Y_molten += random_state.normal(scale=noise, size=Y_molten.size)
     Y = Y_molten.reshape(n_samples)
